#112 refactor(fe): icon객체화 및 적용
