+++ src/frontend/apps/web/src/features/video/model/index.ts
@@ -1,3 +1,5 @@
 export { type HuddleControl } from './huddle-control.type';
-
-export { type HuddleControlAction } from './huddle-controls-reducer';
+export {
+  type HuddleControlAction,
+  huddleControlReducer,
+} from './huddle-controls-reducer';
